tests/mobile-testing/requirements.txt
# Core Dependencies (Required)
robotframework==6.0.1
robotframework-appiumlibrary==2.1.0
Appium-Python-Client>=2.7.1

# Test Data & Configuration (Recommended)
PyYAML==6.0.1                         # YAML test data files
python-dotenv==1.0.0                  # Environment variable management

# Database Integration (Optional)
pymysql==1.1.0                        # MySQL database helper
